+++ qos/evm/errors.go
@@ -1,29 +1,58 @@
 package evm
 
 import (
+	"fmt"
+
 	"github.com/buildwithgrove/path/qos/jsonrpc"
 )
 
-// errResponseEmptyEndpointResponse is a pre-defined JSON-RPC response structure used when an EVM endpoint
-// returns an empty response body. This indicates:
-// - Server misbehavior
-// - Triggers endpoint removal from selection pool
-// - Allows client retry with different endpoint
-var errResponseEmptyEndpointResponse = jsonrpc.GetErrorResponse(
-	jsonrpc.ID{}, // Use request's original ID if present
-	-32000,       // JSON-RPC standard server error code; https://www.jsonrpc.org/historical/json-rpc-2-0.html
-	"Endpoint (data/service node error): Received an empty response. The endpoint will be dropped from the selection pool. Please try again.", // Error Response Message
-	map[string]string{
-		// Marks the error as retryable to allow clients to safely retry their request.
-		"retryable": "true",
-	},
-)
+// newErrResponseInternalErr creates a JSON-RPC error response when an internal error has occurred (e.g. reading HTTP request's body)
+// Marks the error as retryable to allow clients to safely retry their request.
+func newErrResponseInternalErr(err error) jsonrpc.Response {
+	return jsonrpc.GetErrorResponse(
+		jsonrpc.ID{},
+		-32000, // JSON-RPC standard server error code; https://www.jsonrpc.org/historical/json-rpc-2-0.html
+		fmt.Sprintf("internal error: %s", err.Error()), // Error Message
+		map[string]string{
+			"error": err.Error(),
+			// Marks the error as retryable to allow clients to safely retry their request
+			"retryable": "true",
+		},
+	)
+}
 
-// NewEmptyResponse creates a JSON-RPC error response for empty endpoint responses:
+// newErrResponseInvalidRequest returns a JSON-RPC error response for malformed or invalid requests.
+// The error indicates the request cannot be processed due to issues like:
+//   - Failed JSON-RPC deserialization
+//   - Missing required JSON-RPC fields (e.g. `method`)
+//   - Unsupported JSON-RPC method
+//
+// If the request contains a valid JSON-RPC ID, it is included in the error response.
+// The error is marked as permanent since retrying without correcting the request will fail.
+func newErrResponseInvalidRequest(err error, requestID jsonrpc.ID) jsonrpc.Response {
+	return jsonrpc.GetErrorResponse(
+		requestID, // Use request's original ID if present
+		-32000,    // JSON-RPC standard server error code; https://www.jsonrpc.org/historical/json-rpc-2-0.html
+		fmt.Sprintf("invalid request: %s", err.Error()), // Error Message
+		map[string]string{
+			"error": err.Error(),
+			// Indicates this error is permanent - the request must be corrected as retrying will not succeed
+			"retryable": "false",
+		},
+	)
+}
+
+// newErrResponseEmptyEndpointResponse creates a JSON-RPC error response for empty endpoint responses:
 // - Preserves original request ID
 // - Marks error as retryable for safe client retry
-func NewEmptyResponseError(requestID jsonrpc.ID) jsonrpc.Response {
-	response := errResponseEmptyEndpointResponse
-	response.ID = requestID
-	return response
+func newErrResponseEmptyEndpointResponse(requestID jsonrpc.ID) jsonrpc.Response {
+	return jsonrpc.GetErrorResponse(
+		requestID, // Use request's original ID if present
+		-32000,    // JSON-RPC standard server error code; https://www.jsonrpc.org/historical/json-rpc-2-0.html
+		"Endpoint (data/service node error): Received an empty response. The endpoint will be dropped from the selection pool. Please try again.", // Error Response Message
+		map[string]string{
+			// Marks the error as retryable to allow clients to safely retry their request.
+			"retryable": "true",
+		},
+	)
 }
